Predefined Currency converter
This is a calculator to convert from USD to the top 10 populated country currencies. All the errors have been predefined with exceptional handling in order to prevent any further errors during the execution. At the end of conversion, a prompt has been designed to ask the user whether the user wants to continue the conversion or end the program.